Added the sparse matrix multiplication functionality on the GPU so that one may run with a threshold and use the GPU at the same time.
A fourier-transform implementation is still pending. The sparse hack is no longer relevant as the implementation has been rolled back to the current standard way to do sparse matrices in CUDA (while we are waiting for the next major release where NVIDIA claims that the new sparse matrix API will appear on Windows also).