Fixed crash bug with using CUDA. Several pointers where not set to NULL after being freed which led to issues upon subsequent calls.
Also fixed 32-bit precision compliance so now MagTenseMicroMag runs per default on 32 precision with reals to be inline with CUDA and to be able to use MKL's gemv methods. One may use 64bit anywhere in the code, but the compiler now uses 32 bit by default.

Also started adding return the N-tensor mode - both in directly in memory and as a file on the disk. This feature is not done yet, so please don't use it for now.